Fix linting errors in the main code

+++ novelwriter/gui/itemdetails.py
@@ -228,7 +228,7 @@ class GuiItemDetails(QWidget):
             return
 
         self._handle = tHandle
-        iPx = int(round(0.9*SHARED.theme.baseIconHeight))
+        iPx = round(0.9*SHARED.theme.baseIconHeight)
 
         # Label
         # =====
